# docs/contracts/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Cerberus API",
    "description": "Cerberus encrypted vault API. Identity authentication and vault access are separate.",
    "version": "v1"
  },
  "paths": {
...
    "/api/folders/{id}": {
      "delete": {
        "tags": [
          "Folders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Cerberus-Vault-Access",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteFolderCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteFolderCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteFolderCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteFolderOutputDataOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteFolderOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteFolderOutputDataOutput"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "413": {
            "description": "Content Too Large",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Folders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Cerberus-Vault-Access",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFolderCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFolderCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFolderCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateFolderOutputDataOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateFolderOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateFolderOutputDataOutput"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "413": {
            "description": "Content Too Large",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Folders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Cerberus-Vault-Access",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FolderDetailsOutputDataOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderDetailsOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderDetailsOutputDataOutput"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "413": {
            "description": "Content Too Large",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeleteFolderCommand": {
        "required": [
          "expectedRevision"
        ],
        "type": "object",
        "properties": {
          "expectedRevision": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
      "DeleteFolderOutput": {
        "type": "object",
        "properties": {
          "folderId": {
            "type": "string",
            "format": "uuid"
          },
          "trashOperationId": {
            "type": "string",
            "format": "uuid"
          },
          "revision": {
            "type": "integer",
            "format": "int64"
          },
          "serverSequence": {
            "type": "integer",
            "format": "int64"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time"
          },
          "purgeAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "DeleteFolderOutputDataOutput": {
        "type": "object",
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "success": {
            "type": "boolean",
            "readOnly": true
          },
          "data": {
            "$ref": "#/components/schemas/DeleteFolderOutput"
          }
        },
        "additionalProperties": false
      },
...
      "EncryptedEnvelope": {
        "type": "object",
        "properties": {
          "format": {
            "type": "string",
            "nullable": true
          },
          "keyEpoch": {
            "type": "integer",
            "format": "int64"
          },
          "keySalt": {
            "type": "string",
            "nullable": true
          },
          "nonce": {
            "type": "string",
            "nullable": true
          },
          "ciphertext": {
            "type": "string",
            "nullable": true
          },
          "tag": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FolderDetailsOutput": {
        "required": [
          "envelope"
        ],
        "type": "object",
        "properties": {
          "folderId": {
            "type": "string",
            "format": "uuid"
          },
          "revision": {
            "type": "integer",
            "format": "int64"
          },
          "serverSequence": {
            "type": "integer",
            "format": "int64"
          },
          "editedAt": {
            "type": "string",
            "format": "date-time"
          },
          "envelope": {
            "$ref": "#/components/schemas/EncryptedEnvelope"
          },
          "profileIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "parentFolderId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "collectionIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FolderDetailsOutputDataOutput": {
        "type": "object",
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "success": {
            "type": "boolean",
            "readOnly": true
          },
          "data": {
            "$ref": "#/components/schemas/FolderDetailsOutput"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
      "ProcessOutput": {
        "type": "object",
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "success": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateFolderCommand": {
        "required": [
          "editedAt",
          "envelope",
          "expectedRevision"
        ],
        "type": "object",
        "properties": {
          "expectedRevision": {
            "type": "integer",
            "format": "int64"
          },
          "envelope": {
            "$ref": "#/components/schemas/EncryptedEnvelope"
          },
          "editedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "UpdateFolderOutput": {
        "type": "object",
        "properties": {
          "folderId": {
            "type": "string",
            "format": "uuid"
          },
          "revision": {
            "type": "integer",
            "format": "int64"
          },
          "serverSequence": {
            "type": "integer",
            "format": "int64"
          },
          "editedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "UpdateFolderOutputDataOutput": {
        "type": "object",
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "success": {
            "type": "boolean",
            "readOnly": true
          },
          "data": {
            "$ref": "#/components/schemas/UpdateFolderOutput"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}